+++ impl_core/src/test/java/io/opencensus/implcore/stats/CurrentStatsStateTest.java
@@ -40,9 +40,9 @@ public final class CurrentStatsStateTest {
   public void setState() {
     CurrentStatsState state = new CurrentStatsState();
     assertThat(state.set(StatsCollectionState.DISABLED)).isTrue();
-    assertThat(state.getInternal()).isEqualTo(StatsCollectionState.DISABLED);
+    assertThat(state.get()).isEqualTo(StatsCollectionState.DISABLED);
     assertThat(state.set(StatsCollectionState.ENABLED)).isTrue();
-    assertThat(state.getInternal()).isEqualTo(StatsCollectionState.ENABLED);
+    assertThat(state.get()).isEqualTo(StatsCollectionState.ENABLED);
     assertThat(state.set(StatsCollectionState.ENABLED)).isFalse();
   }
 
@@ -52,14 +52,5 @@ public final class CurrentStatsStateTest {
     thrown.expect(NullPointerException.class);
     thrown.expectMessage("state");
     state.set(null);
-  }
-
-  @Test
-  public void preventSettingStateAfterReadingState() {
-    CurrentStatsState state = new CurrentStatsState();
-    state.get();
-    thrown.expect(IllegalStateException.class);
-    thrown.expectMessage("State was already read, cannot set state.");
-    state.set(StatsCollectionState.DISABLED);
   }
 }